refactor: better project structure
