--- src/libs/mod.rs
@@ -1,4 +1,0 @@
-pub mod error;
-pub mod hst;
-pub mod ins;
-pub mod vrs;
