+++ src/task/inv.rs
@@ -1,13 +1,11 @@
 use std::ops::{Deref, DerefMut};
 
-use genin::libs::{
-    error::{ConfigError, TaskError},
-    ins::{is_false, Instance, Role, Type},
-};
 use indexmap::IndexMap;
 use log::{debug, trace};
 use serde::{Deserialize, Serialize};
 use serde_yaml::Value;
+use crate::error::{ConfigError, TaskError};
+use crate::task::ins::{Instance, is_false, Role, Type};
 use super::cluster::scheme::Scheme;
 
 #[derive(Serialize, Deserialize)]
